+++ cayman/__init__.py
@@ -1,3 +1,3 @@
 """ module docstring """
-__version__ = "0.11.0"
+__version__ = "0.11.1"
 __toolname__ = "cayman"