+++ CHANGELOG.md
@@ -4,7 +4,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [v0.11.0] - 2026-03-
+## [v0.11.1] - 2026-04-21
+This update allows the user to supply an alignment file as input instead of reads directly
+[v0.11.1]: https://github.com/zellerlab/cayman/compare/v0.11.0...v0.11.1
+
+### Added
+- `--sam` optional input flag added to pass SAM alignment file
+- `--bam` optional input flag added to pass BAM alignment file
+- `--import_readcounts` <int> and `-unmarked_orphans` flags added to correct counting from alignment file
+- `--keep_alignment_file` <path> added which will save the SAM alignment file in case reads were passed as input
+
+### Breaking
+- `--cpus_for_alignment` flag got renamed `--threads` in the profile subparser
+- `bwa_index` is no longer a positional argument but now is only required with read input. pass it after the `--bwa_index` flag
+
+
+## [v0.11.0] - 2026-03-20
 This is a purely performance driven update. IO remains exactly the same - just 4-5x faster
 
 ### Added